requirements.txt
Flask==2.0.1

pip~=21.1.3
wheel~=0.36.2
lxml~=4.6.2
setuptools~=57.0.0
python-dotenv~=0.15.0
scikit-learn~=1.0
numpy~=1.21.2
scikit-learn-extra~=0.2.0